requirements.txt
boto3
bcrypt
pyDNS
validate_email
PyJWT
